+++ tests/conftest.py
@@ -3,6 +3,8 @@ import platform
 
 import numpy as np
 import pytest
+import shutil as sh
+from pathlib import Path
 from mne.utils import run_subprocess
 
 from mne_hfo.simulate import simulate_hfo, simulate_spike
@@ -36,6 +38,17 @@ def _bids_validate():
         run_subprocess(cmd, shell=shell)
 
     return _validate
+
+
+@pytest.fixture(scope='function')
+def test_bids_root(tmpdir):
+    """Temporary BIDS dataset.
+
+    Copies over dataset in ``data/`` to temporary directory.
+    """
+    data_path = Path('data')
+    sh.copytree(data_path, tmpdir, dirs_exist_ok=True)
+    return tmpdir
 
 
 @pytest.fixture(scope="module")